urbit: consolidate client logic further and handle logged out case
